Add 'Preview' to context menu
