+++ frontend/src/core/app/helpers.js
@@ -358,6 +358,92 @@ define(function(require){
         }
       },
 
+      previewProject: function(course, courseId) {
+        var self = this;
+
+        if (this.validateCourseContent(course) && !Origin.helper.isPreviewPending) {
+          Origin.helper.isPreviewPending = true;
+          $('.navigation-loading-indicator').removeClass('display-none');
+          $('.editor-common-sidebar-preview-inner').addClass('display-none');
+          $('.editor-common-sidebar-previewing').removeClass('display-none');
+
+          $.ajax({
+            method: 'get',
+            url: '/api/output/' + Origin.constants.outputPlugin + '/preview/' + courseId,
+            success: function (jqXHR, textStatus, errorThrown) {
+              if (jqXHR.success) {
+                if (jqXHR.payload && typeof(jqXHR.payload.pollUrl) != 'undefined' && jqXHR.payload.pollUrl != '') {
+                  // Ping the remote URL to check if the job has been completed
+                  self.updatePreviewProgress(jqXHR.payload.pollUrl);
+                } else {
+                  self.launchCoursePreview(courseId);
+                  self.resetPreviewProgress();
+                }
+              } else {
+                self.resetPreviewProgress();
+                Origin.Notify.alert({
+                  type: 'error',
+                  text: window.polyglot.t('app.errorgeneratingpreview')
+                });
+              }
+            },
+            error: function (jqXHR, textStatus, errorThrown) {
+              self.resetPreviewProgress();
+              Origin.Notify.alert({
+                type: 'error',
+                text: window.polyglot.t('app.errorgeneric')
+              });
+            }
+          });
+        }
+      },
+
+      updatePreviewProgress: function(url) {
+        var self = this;
+
+        var pollUrl = function() {
+          $.ajax({
+            method: 'get',
+            url: url,
+            success: function(jqXHR, textStatus, errorThrown) {
+              if (jqXHR.progress == "100") {
+                clearInterval(pollId);
+                self.launchCoursePreview();
+                self.resetPreviewProgress();
+              } else {
+                $('.navigation-loading-progress').animate({ width: jqXHR.progress + '%' }, 1000);
+              }
+            },
+            error: function(jqXHR, textStatus, errorThrown) {
+              clearInterval(pollId);
+              self.resetPreviewProgress();
+
+              Origin.Notify.alert({
+                type: 'error',
+                text: errorThrown
+              });
+            }
+          });
+          };
+
+          // Check for updated progress every 3 seconds
+          var pollId = setInterval(pollUrl, 3000);
+      },
+
+      launchCoursePreview: function(courseId) {
+        var tenantId = Origin.sessionModel.get('tenantId');
+
+        window.open('/preview/' + tenantId + '/' + courseId + '/', 'preview');
+      },
+
+      resetPreviewProgress: function() {
+        $('.navigation-loading-progress').css('width', 0).stop();
+        $('.editor-common-sidebar-preview-inner').removeClass('display-none');
+        $('.editor-common-sidebar-previewing').addClass('display-none');
+        $('.navigation-loading-indicator').addClass('display-none');
+        Origin.helper.isPreviewPending = false;
+      },
+
       isValidEmail: function(value) {
         var regEx = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
         if (value.length === 0 || !regEx.test(value)) {
