+++ frontend/src/core/project/views/projectView.js
@@ -3,6 +3,7 @@ define(function(require){
 
   var Backbone = require('backbone');
   var Handlebars = require('handlebars');
+  var helpers = require('coreJS/app/helpers');
   var OriginView = require('coreJS/app/views/originView');
   var Origin = require('coreJS/app/origin');
 
@@ -33,6 +34,7 @@ define(function(require){
       this.listenTo(Origin, 'dashboard:projectView:itemSelected', this.deselectItem);
       this.listenTo(Origin, 'dashboard:dashboardView:deselectItem', this.deselectItem);
 
+      this.on('contextMenu:course:openPreview', this.openPreview);
       this.on('contextMenu:course:editSettings', this.editProjectSettings);
       this.on('contextMenu:course:edit', this.editProject);
       this.on('contextMenu:course:delete', this.deleteProjectPrompt);
@@ -46,6 +48,14 @@ define(function(require){
       e.preventDefault();
 
       Origin.trigger('contextMenu:open', this, e);
+    },
+
+    openPreview: function(event) {
+      if (event) {
+          event.preventDefault();
+      }
+
+      helpers.previewProject(this.model, this.model.get('_id'));
     },
 
     editProjectSettings: function(event) {
